+++ internal/lazy/scripting.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"sync"
 
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/db"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/database"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/errors"
 	builder "github.com/ydb-platform/ydb-go-sdk/v3/internal/scripting"
 	"github.com/ydb-platform/ydb-go-sdk/v3/retry"
@@ -15,7 +15,7 @@ import (
 )
 
 type lazyScripting struct {
-	db      db.Connection
+	db      database.Connection
 	options []config.Option
 	client  scripting.Client
 	m       sync.Mutex
@@ -76,7 +76,7 @@ func (s *lazyScripting) Close(ctx context.Context) (err error) {
 	return nil
 }
 
-func Scripting(db db.Connection, options []config.Option) scripting.Client {
+func Scripting(db database.Connection, options []config.Option) scripting.Client {
 	return &lazyScripting{
 		db:      db,
 		options: options,