+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Renamed package `internal/db` to `internal/database` for exclude collisions with variable name `db`
+
 ## v3.18.0
 * Added `go1.18` to test matrix
 * Added `ydb.WithOperationTimeout` and `ydb.WithOperationCancelAfter` context modifiers